requirements.txt
python-decouple==3.8
jira==3.5.2
pandas==2.1.3